adjust: correction in bugs start

+++ database/migrations/2026_04_14_021127_add_course_shift_id_to_enrollments_table.php
@@ -11,13 +11,15 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('enrollments', function (Blueprint $table) {
-            $table->foreignId('course_shift_id')
-                ->nullable()
-                ->after('course_id')
-                ->constrained('course_shifts')
-                ->nullOnDelete();
-        });
+        if (!Schema::hasColumn('enrollments', 'course_shift_id')) {
+            Schema::table('enrollments', function (Blueprint $table) {
+                $table->foreignId('course_shift_id')
+                    ->nullable()
+                    ->after('course_id')
+                    ->constrained('course_shifts')
+                    ->nullOnDelete();
+            });
+        }
     }
 
     /**
@@ -26,10 +28,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('enrollments', function (Blueprint $table) {
-            Schema::table('enrollments', function (Blueprint $table) {
+            if (Schema::hasColumn('enrollments', 'course_shift_id')) {
                 $table->dropForeign(['course_shift_id']);
                 $table->dropColumn('course_shift_id');
-            });
+            }
         });
     }
 };
